refactor(router-cli): domain-separated folder structure per ADR-0005
Colocate CLI definitions, models, and handlers per resource domain.
Each resource (proxy, domain, gwnode, gateway, user, cert, sync)
gets its own folder with mod.rs and models.rs. Shared types move
to common.rs. cli.rs shrinks to top-level Commands enum only.

Adding a new resource now means creating one folder with no
changes to existing files.
