+++ docs/adr/README.md
@@ -12,6 +12,7 @@ ADRs document significant architectural decisions along with their context, rati
 | [0002](0002-cli-architecture-redesign.md) | CLI Architecture Redesign | 2026-03-29 |
 | [0003](0003-cli-control-panel.md) | CLI Control Panel | 2026-03-29 |
 | [0004](0004-cli-live-monitoring.md) | CLI Live Monitoring | 2026-03-29 |
+| [0005](0005-cli-domain-separation.md) | CLI Domain Separation | 2026-03-29 |
 
 ## How to Contribute
